Rename : 토글 스위치 이름 변경
