+++ src/components/commons/Switch/Switch.stories.js
@@ -1,27 +1,27 @@
 import React from "react";
 
-import Toggle from "./";
+import Switch from ".";
 import { Background } from "@/foundations";
 
 export default {
-  title: "Components/Toggle",
-  component: Toggle,
+  title: "Components/Switch",
+  component: Switch,
 };
 
 export const Default = (args) => (
   <Background {...args}>
-    <Toggle {...args} />
+    <Switch {...args} />
   </Background>
 );
 
 export const AllTypes = () => (
   <>
     <Background theme="light">
-      <Toggle theme="light" />
+      <Switch theme="light" />
     </Background>
     <br />
     <Background theme="dark">
-      <Toggle theme="dark" />
+      <Switch theme="dark" />
     </Background>
   </>
 );
