--- src/components/commons/Toggle/index.jsx
@@ -1,1 +1,0 @@
-export { default } from "./Toggle";
