+++ src/components/commons/index.js
@@ -5,4 +5,4 @@ export { default as Button } from "./Button";
 export { default as ImageBox } from "./ImageBox";
 export { default as Input } from "./Input";
 export { default as SideBarItem } from "./SideBarItem";
-export { default as Toggle } from "./Toggle";
+export { default as Switch } from "./Switch";
